adding GameweekHelperTests

+++ src/main/scala/gameweek/GameweekSchema.scala
@@ -6,7 +6,7 @@ object GameweekSchema {
 
   val gameweekStruct = new StructType(
     Array[StructField](
-      StructField("transfers_balance", DataTypes.IntegerType, true, Metadata.empty),
+      StructField("transfers_balance", DataTypes.LongType, true, Metadata.empty),
       StructField("opponent_team", DataTypes.StringType, true, Metadata.empty),
       StructField("bonus", DataTypes.IntegerType, true, Metadata.empty),
       StructField("own_goals", DataTypes.IntegerType, true, Metadata.empty),
@@ -19,10 +19,10 @@ object GameweekSchema {
       StructField("influence", DataTypes.DoubleType, true, Metadata.empty),
       StructField("saves", DataTypes.IntegerType, true, Metadata.empty),
       StructField("assists", DataTypes.IntegerType, true, Metadata.empty),
-      StructField("transfers_in", DataTypes.IntegerType, true, Metadata.empty),
+      StructField("transfers_in", DataTypes.LongType, true, Metadata.empty),
       StructField("creativity", DataTypes.DoubleType, true, Metadata.empty),
       StructField("value", DataTypes.IntegerType, true, Metadata.empty),
-      StructField("selected", DataTypes.IntegerType, true, Metadata.empty),
+      StructField("selected", DataTypes.LongType, true, Metadata.empty),
       StructField("element", DataTypes.IntegerType, true, Metadata.empty),
       StructField("goals_scored", DataTypes.IntegerType, true, Metadata.empty),
       StructField("bps", DataTypes.IntegerType, true, Metadata.empty),
@@ -32,7 +32,7 @@ object GameweekSchema {
       StructField("yellow_cards", DataTypes.IntegerType, true, Metadata.empty),
       StructField("team", DataTypes.StringType, true, Metadata.empty),
       StructField("fixture", DataTypes.IntegerType, true, Metadata.empty),
-      StructField("transfers_out", DataTypes.IntegerType, true, Metadata.empty),
+      StructField("transfers_out", DataTypes.LongType, true, Metadata.empty),
       StructField("round", DataTypes.IntegerType, true, Metadata.empty),
       StructField("name", DataTypes.StringType, true, Metadata.empty),
       StructField("threat", DataTypes.DoubleType, true, Metadata.empty),
